Update dependencies and code cleanups
Update Vue to ^2.6.14 and all other dependencies to their latest (or
newest when not possible).

Fix some warnings in Vue templates.

Add gemmaConfig.js to configure a base URL for Gemma. Use it for the API
and links to EEs and platforms.

Remove yarn.lock in favour of package-lock.json.