Specify the order of categories in the annotationSelector
Make an array in gemma.js specifying the order of the categories in the annotationSelector component (annotationSelectorOrderArray).

Import the new annotationSelectorOrderArray into AnnotationSelector.vue. Replace the alphabetical sort function in AnnotationSelector.vue with a new sort function that displays the categories in the annotationSelector component according to the order set in annotationSelectorOrderArray.