Various fixes
Sort annotations by number of datasets instead of entropy. This way
options are never shuffled when the total number of datasets displayed
changed.

Add a strikethrough to quickly identify terms in the exclusion list. It
does not cover inferred terms however.

Use getCategoryId() consistently.

Add some bottom margins to the SearchSettings items.

Add a heading to the taxon selector.